refactor: moved the Block into separate components.
Moved the Block into separate components.
