+++ my-app/src/style/style.js
@@ -1,4 +1,19 @@
 const style = {
+    content: {
+        width: 355,
+        margin: "auto",
+        marginTop: 30,
+    },
+    head: {
+        display: "flex",
+    },
+    logo: {
+        fontFamily: "sans-serif",
+        flex: 1,
+        fontWeight: "700",
+        fontSize: 60,
+        color: "#776e65",
+    },
     blockStyle: {
         height: 80,
         width: 80,
@@ -11,6 +26,10 @@ const style = {
         fontWeight: "800",
         color: "white",
     },
+    newGameButtonBlock:{
+        flex: 1,
+        marginTop: "auto",
+    },
     newGameButton: {
         padding: 10,
         background: "#846F5B",
@@ -21,6 +40,19 @@ const style = {
         marginLeft: "auto",
         marginBottom: "auto",
         cursor: "pointer",
+    },
+    playingАield:{
+        background: "#AD9D8F",
+        width: "max-content",
+        height: "max-content",
+        margin: "auto",
+        padding: 5,
+        borderRadius: 5,
+        marginTop: 10,
+        position: "relative",
+    },
+    tryAgainButtonBlock:{
+
     },
     tryAgainButton: {
         padding: 10,
@@ -45,6 +77,12 @@ const style = {
         justifyContent: "center",
         alignItems: "center",
     },
+    gameOverOverlayMessage:{
+        fontSize: 30,
+        fontFamily: "sans-serif",
+        fontWeight: "900",
+        color: "#776E65",
+    }
 };
 
 export default style;
